Add ExtractConfition class
