+++ src/main/java/jp/co/future/uroborosql/fluent/SqlEntityQuery.java
@@ -19,14 +19,14 @@ import jp.co.future.uroborosql.exception.UroborosqlRuntimeException;
  * @param <E> Entity型
  * @author ota
  */
-public interface SqlEntityQuery<E> extends SqlFluent<SqlEntityQuery<E>> {
+public interface SqlEntityQuery<E> extends ExtractionCondition<SqlEntityQuery<E>> {
 	/**
 	 * ORDER BY句の順序を表すEnum
 	 */
 	public enum Order {
 		ASCENDING("ASC"), DESCENDING("DESC");
 
-		private final String alias;
+		private String alias;
 
 		Order(final String alias) {
 			this.alias = alias;
@@ -72,180 +72,6 @@ public interface SqlEntityQuery<E> extends SqlFluent<SqlEntityQuery<E>> {
 	 * @return 検索結果を順次取得するStream.
 	 */
 	Stream<E> stream();
-
-	/**
-	 * Where句に equal 条件を追加する
-	 * @param col bind column name
-	 * @param value 値
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> equal(String col, Object value);
-
-	/**
-	 * Where句に not equal 条件を追加する
-	 * @param col bind column name
-	 * @param value 値
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> notEqual(String col, Object value);
-
-	/**
-	 * Where句に greater than 条件を追加する
-	 * @param col bind column name
-	 * @param value 値
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> greaterThan(String col, Object value);
-
-	/**
-	 * Where句に less than 条件を追加する
-	 * @param col bind column name
-	 * @param value 値
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> lessThan(String col, Object value);
-
-	/**
-	 * Where句に greater equal 条件を追加する
-	 * @param col bind column name
-	 * @param value 値
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> greaterEqual(String col, Object value);
-
-	/**
-	 * Where句に less equal 条件を追加する
-	 * @param col bind column name
-	 * @param value 値
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> lessEqual(String col, Object value);
-
-	/**
-	 * Where句に in 条件を追加する
-	 * @param col bind column name
-	 * @param values 値の配列
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> in(String col, Object... values);
-
-	/**
-	 * Where句に in 条件を追加する
-	 * @param col bind column name
-	 * @param valueList 値の集合
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> in(String col, Iterable<?> valueList);
-
-	/**
-	 * Where句に not in 条件を追加する
-	 * @param col bind column name
-	 * @param values 値の配列
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> notIn(String col, Object... values);
-
-	/**
-	 * Where句に not in 条件を追加する
-	 * @param col bind column name
-	 * @param valueList 値の集合
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> notIn(String col, Iterable<?> valueList);
-
-	/**
-	 * Where句に like 条件を追加する。検索文字列はエスケープしない
-	 * @param col bind column name
-	 * @param searchValue 検索文字列
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> like(String col, CharSequence searchValue);
-
-	/**
-	 * Where句に前方一致条件を追加する。検索文字列は内部でエスケープされる。
-	 * @param col bind column name
-	 * @param searchValue 検索文字列
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> startsWith(String col, CharSequence searchValue);
-
-	/**
-	 * Where句に後方一致条件を追加する。検索文字列は内部でエスケープされる。
-	 * @param col bind column name
-	 * @param searchValue 検索文字列
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> endsWith(String col, CharSequence searchValue);
-
-	/**
-	 * Where句に 部分一致条件を追加する。検索文字列は内部でエスケープされる。
-	 * @param col bind column name
-	 * @param searchValue 検索文字列
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> contains(String col, CharSequence searchValue);
-
-	/**
-	 * Where句に not like 条件を追加する
-	 * @param col bind column name
-	 * @param searchValue 検索文字列
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> notLike(String col, CharSequence searchValue);
-
-	/**
-	 * Where句に前方一致条件の否定を追加する。検索文字列は内部でエスケープされる。
-	 * @param col bind column name
-	 * @param searchValue 検索文字列
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> notStartsWith(String col, CharSequence searchValue);
-
-	/**
-	 * Where句に後方一致条件の否定を追加する。検索文字列は内部でエスケープされる。
-	 * @param col bind column name
-	 * @param searchValue 検索文字列
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> notEndsWith(String col, CharSequence searchValue);
-
-	/**
-	 * Where句に 部分一致条件の否定を追加する。検索文字列は内部でエスケープされる。
-	 * @param col bind column name
-	 * @param searchValue 検索文字列
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> notContains(String col, CharSequence searchValue);
-
-	/**
-	 * Where句に between 条件を追加する
-	 * @param col bind column name
-	 * @param fromValue 開始値
-	 * @param toValue 終了値
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> between(String col, Object fromValue, Object toValue);
-
-	/**
-	 * Where句に is null 条件を追加する
-	 * @param col bind column name
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> isNull(String col);
-
-	/**
-	 * Where句に is not null 条件を追加する
-	 * @param col bind column name
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> isNotNull(String col);
-
-	/**
-	 * Where句に 素の文字列指定で 条件を追加する。複数回呼び出した場合はそれぞれの条件をANDで結合する。
-	 * @param rawString Where句に出力する条件式
-	 * @return SqlEntityQuery
-	 */
-	SqlEntityQuery<E> where(CharSequence rawString);
 
 	/**
 	 * ソート条件を指定（昇順）
